pyproject.toml
[project]
name = "compliance-checks"
version = "0.1.0"
description = "Compliance checks for Google Cloud IAM policies"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "google-api-python-client>=2.108.0",
    "google-cloud-asset>=3.24.0",
    "google-auth>=2.23.4",
    "pydantic>=2.5.0",
    "google-cloud-resource-manager>=1.14.2",
    "structlog>=23.2.0",
    "google-cloud-firestore>=2.13.1",
    "tinydb>=4.8.0",
]
